[rename]: 폴더명 변경 (#6)
ProductItem -> ColumnProductItem
