+++ src/components/ColumnProductItem/Thumbnail.module.scss
@@ -1,5 +1,5 @@
-@import '@styles/_variable.module';
-@import '@styles/_mixins.module';
+@import '@styles/mixins.module';
+@import '@styles/variable.module';
 
 $round: 8px;
