+++ src/components/ColumnProductItem/Thumbnail.tsx
