+++ src/components/ColumnProductItem/index.module.scss
