+++ src/components/index.ts
@@ -1,2 +1,2 @@
 export { default as SearchResult } from './SearchResult';
-export { default as ColumnProductItem } from './ProductItem';
+export { default as ColumnProductItem } from './ColumnProductItem';
